+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## 2.26.1
+
+### Refactor
+
+- refactor mybatis autofill.
+
+### Chore
+
+#### Scaffold
+
+- Update scaffold default radpVersion to `2.26.1`.
+
 ## 2.26
 
 ### Feature